One more package move for CAM.
I really think I'm getting close.
