+++ src/main/java/com/marklogic/mule/connector/api/error/ErrorType.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.marklogic.mule.connector.internal.error;
+package com.marklogic.mule.connector.api.error;
 
 import org.mule.runtime.extension.api.error.ErrorTypeDefinition;
