+++ src/main/java/com/marklogic/mule/connector/internal/operation/ReadPagingProvider.java
@@ -29,7 +29,7 @@ import com.marklogic.client.io.JacksonParserHandle;
 import com.marklogic.client.query.QueryDefinition;
 import com.marklogic.mule.connector.internal.Utilities;
 import com.marklogic.mule.connector.api.types.DocumentAttributes;
-import com.marklogic.mule.connector.internal.error.ErrorType;
+import com.marklogic.mule.connector.api.error.ErrorType;
 import org.mule.runtime.api.metadata.MediaType;
 import org.mule.runtime.extension.api.exception.ModuleException;
 import org.mule.runtime.extension.api.runtime.operation.Result;
